Веб сервер

+++ agent/internal/computing_service/task.go
@@ -2,9 +2,9 @@ package computingservice
 
 type Task struct {
 	id             int
-	arg1           int
-	arg2           int
+	arg1           float64
+	arg2           float64
 	operation_time int
 	operation      string
-	result         int
+	result         float64
 }
