+++ server/internal/expression_service/schema.go
@@ -5,20 +5,20 @@ type CalculateResponseSchema struct {
 }
 
 type ExpressionResponseSchema struct {
-	Id     int  `json:"id"`
-	Status bool `json:"status"`
-	Result int  `json:"result"`
+	Id     int     `json:"id"`
+	Status bool    `json:"status"`
+	Result float64 `json:"result"`
 }
 
 type TaskSchema struct {
-	Id             int    `json:"id"`
-	Arg1           int    `json:"arg1"`
-	Arg2           int    `json:"arg2"`
-	Operation      string `json:"operation"`
-	Operation_time int    `json:"operation_time"`
+	Id             int     `json:"id"`
+	Arg1           float64 `json:"arg1"`
+	Arg2           float64 `json:"arg2"`
+	Operation      string  `json:"operation"`
+	Operation_time int     `json:"operation_time"`
 }
 
 type TaskSchemaRequest struct {
-	Id     int `json:"id"`
-	Result int `json:"result"`
+	Id     int     `json:"id"`
+	Result float64 `json:"result"`
 }
